web-agent/web-app/requirements.txt
Flask==2.2.2
python-dotenv==0.20.0
pydantic==1.10.2
asyncio==3.4.3